Separate custom_sampler's python file
