+++ netket/sampler/__init__.py
@@ -3,4 +3,5 @@ from .metropolis_hastings import *
 from .metropolis_local import MetropolisLocal, MetropolisLocalPt
 from .metropolis_exchange import MetropolisExchange, MetropolisExchangePt
 from .metropolis_hamiltonian import MetropolisHamiltonian, MetropolisHamiltonianPt
+from .custom_sampler import CustomSampler, CustomSamplerPt
 from .exact_sampler import *
